requirements.txt
pytz
ed25519
# recommended, but not necessary
python-rapidjson
